Change generate_hash
Deal with the case (as was happening when I ran this (on a folder which was a synced Sharepoint online document library) where there was a file which I had no read-permissions on. The code shouldn't raise an exception here, as a "can't open" file should just be ignored for the purposes of matching as a dupe.